+++ package.json
@@ -41,6 +41,7 @@
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-loading-skeleton": "^2.1.1",
+    "react-tagsinput": "^3.19.0",
     "reactstrap": "^8.5.1",
     "sass": "^1.26.10",
     "stylelint": "^13.6.1",
